added alert to  menu and product page

+++ admin/assets/js/menu.js
@@ -1,7 +1,5 @@
 const BASE_URL = "http://localhost:8080/";
 
-
-
 const tBody = document.querySelector("tbody");
 const searchInput = document.querySelector("#simple-search");
 
@@ -14,14 +12,16 @@ function createCard(arr) {
   arr.forEach((element, index) => {
     tBody.innerHTML += `
 <tr>
-<td>${index+1}</td>
+<td>${index + 1}</td>
 <td><img src=${element.image} width="70px"></td>
 <td>${element.name}</td>
 <td class="description">${element.description}</td>
 <td>${element.price}</td>
 <td>${element.ingredient}</td>
 <td class="process">
-<a href="#" onclick=deleteFunc(${element.id})><i class="fa-solid fa-trash"></i></a>
+<a href="#" onclick=deleteFunc(${
+      element.id
+    })><i class="fa-solid fa-trash"></i></a>
 <a href="menuForm.html?id=${element.id}"><i class="fa-solid fa-pen"></i></a>
 </td>
 </tr>`;
@@ -46,6 +46,23 @@ searchInput.addEventListener("input", function (e) {
   createCard(copyArr);
 });
 
-async function deleteFunc(id) {
-  await axios.delete(`${BASE_URL}menu/${id}`);
+// Delete
+function deleteFunc(id) {
+  swal({
+    title: "Are you sure?",
+    text: "Once deleted, you will not be able to recover this imaginary file!",
+    icon: "warning",
+    buttons: true,
+    dangerMode: true,
+  }).then((willDelete) => {
+    if (willDelete) {
+      axios.delete(`${BASE_URL}menu/${id}`);
+
+      swal("Poof! Your imaginary file has been deleted!", {
+        icon: "success",
+      });
+    } else {
+      swal("Your imaginary file is safe!");
+    }
+  });
 }
